fix(desktop): align the setup-failure panel with its live sibling
Review follow-ups on the height bound.

`overflow-y: visible` on the output pre was inert: `__pre` sets
`overflow-x: auto`, which forces the other axis to compute to `auto`, so the
declaration said the opposite of the used value. Worse, dropping the output's
own 320px viewport had moved its horizontal scrollbar to the bottom of the full
output, outside the body's viewport, on exactly the wide-line output that needs
it. `.launchpad-pending__output pre` — the live setup panel rendering the same
data in the same pane — wraps instead, so do that here: no sideways scroll to
reach, and the two panels stop diverging.

The `<details>` summary carries the exit code and the collapse control, and it
was scrolling out of the panel it belongs to. Pin it, the way the live panel
pins its section header.

Cap at `min(360px, 40vh)` rather than a bare `40vh`, matching
`.launchpad-pending__command`: a 1400px-tall window was handing the panel 560px
of body for no reason (the transcript goes from 570px to 850px there).

Scroll the body back to the top when a Continue failure appears — the error
renders in the heading block, which is now inside the scroller, so an operator
who had scrolled down to read the output got a re-enabled button and no visible
message.

Tests: `ruleBody` was on its third copy in this directory, so it moves to a
shared `css-rule-body` module the other two contract tests now use. The
max-height assertion matched only a bare integer `vh`, which would have
rejected the `min(px, vh)` cap it exists to protect.
